Create health reminder
this is a program named health reminder. this program reminds health tasks like eye exercise, physical exercise, and drinking water to a person who works 9 - 5 a day continuously. it plays particular audio for a particular task. so that if the person is not watching the output screen of the program it will still remind him about his health tasks.